created message protocol

+++ libfactorization.h
@@ -1,6 +1,6 @@
 #ifndef _LIBFACTORIZATION_
 #define _LIBFACTORIZATION_
 
-int factorize(int num, int **factors, int *factors_len);
+int factorize(unsigned short num, unsigned short **factors, int *factors_len);
 
 #endif
